Moved Firebase setup to network folder
