+++ src/network/config.ts
@@ -1,0 +1,12 @@
+const firebaseConfig = {
+  apiKey: "AIzaSyAG4FPlVFT20lLKjErXY2LH0AICB070gsI",
+  authDomain: "neighbor-to-neighbor-da5ce.firebaseapp.com",
+  databaseURL: "https://neighbor-to-neighbor-da5ce.firebaseio.com",
+  projectId: "neighbor-to-neighbor-da5ce",
+  storageBucket: "neighbor-to-neighbor-da5ce.appspot.com",
+  messagingSenderId: "682085165",
+  appId: "1:682085165:web:8168de7754befe55",
+};
+// for reference: https://stackoverflow.com/questions/37482366/is-it-safe-to-expose-firebase-apikey-to-the-public#37484053
+
+export default firebaseConfig;
